Add files via upload
Sustainable development focuses on conscientious use of resources to protect the
environment, economy, and society. On campuses, resources are wasted where lights and
fans are left open in empty rooms. With increasing technology, especially in the sector of
the Internet of Things (IoT), smart solutions are available to better utilize energy. IoTbased automation enables smart regulation of electrical appliances based on occupancy,
thus limiting wastage and encouraging sustainability.
The system proposed will perform the following operations:
· Identify the presence of a human in a room.
· Identify if the devices in a room are ON or OFF.
· Manage the devices by turning them ON/OFF according to human detection
· Ongoing monitoring of a space for entry/exit by any human being